model_monitoring/requirements.txt
streamlit==1.45.1
pandas==2.2.3
psycopg2-binary==2.9.9
minio==7.2.0
scikit-learn==1.4.0
plotly==5.18.0
python-dotenv==1.0.0
numpy==1.26.3